experimental version of safety-shield
