--- safety_shield/src/safety_shield_alternative.cc
@@ -1,1146 +1,0 @@
-#include "safety_shield/safety_shield.h"
-
-
-namespace safety_shield {
-
-SafetyShield::SafetyShield():
-        max_s_stop_(0),
-        v_max_allowed_({0, 0, 0}),
-        a_max_allowed_({0, 0, 0}),
-        j_max_allowed_({0, 0, 0}),
-        a_max_ltt_({0, 0, 0}),
-        j_max_ltt_({0, 0, 0})
-{
-    spdlog::info("Safety shield created.");
-}
-
-SafetyShield::SafetyShield(bool activate_shield,
-                           int nb_joints,
-                           double sample_time,
-                           double max_s_stop,
-                           const std::vector<double> &v_max_allowed,
-                           const std::vector<double> &a_max_allowed,
-                           const std::vector<double> &j_max_allowed,
-                           const std::vector<double> &a_max_path,
-                           const std::vector<double> &j_max_path,
-                           const LongTermTraj &long_term_trajectory,
-                           RobotReach* robot_reach,
-                           HumanReach* human_reach,
-                           Verify* verify):
-        activate_shield_(activate_shield),
-        nb_joints_(nb_joints),
-        max_s_stop_(max_s_stop),
-        v_max_allowed_(v_max_allowed),
-        a_max_allowed_(a_max_allowed),
-        j_max_allowed_(j_max_allowed),
-        a_max_ltt_(a_max_path),
-        j_max_ltt_(j_max_path),
-        sample_time_(sample_time),
-        path_s_(0),
-        path_s_discrete_(0),
-        long_term_trajectory_(long_term_trajectory),
-        robot_reach_(robot_reach),
-        human_reach_(human_reach),
-        verify_(verify)
-{
-    sliding_window_k_ = (int) std::floor(max_s_stop_/sample_time_);
-    std::vector<double> prev_dq;
-    for(int i = 0; i < 6; i++) {
-        prev_dq.push_back(0.0);
-        alpha_i_.push_back(1.0);
-    }
-    alpha_i_.push_back(1.0);
-
-    is_safe_ = !activate_shield_;
-    is_static_safe_ = is_safe_;
-    is_PFL_safe_ = is_safe_;
-    if(safety_method_ == STANDARD) {
-        computesPotentialTrajectory(is_safe_, prev_dq);
-        next_motion_ = determineNextMotion(is_safe_);
-    } else {
-        computesPotentialTrajectoryForPFL(is_static_safe_, is_PFL_safe_, prev_dq, nullptr, nullptr);
-        next_motion_ = determineNextMotionForPFL(is_static_safe_, is_PFL_safe_);
-    }
-
-    std::vector<double> q_min(nb_joints, -3.141);
-    std::vector<double> q_max(nb_joints, -3.141);
-    ltp_ = long_term_planner::LongTermPlanner(nb_joints, sample_time, q_min, q_max, v_max_allowed, a_max_path, j_max_path);
-    spdlog::info("Safety shield created.");
-}
-
-SafetyShield::SafetyShield(bool activate_shield,
-                           double sample_time,
-                           std::string trajectory_config_file,
-                           std::string robot_config_file,
-                           std::string mocap_config_file,
-                           double init_x,
-                           double init_y,
-                           double init_z,
-                           double init_roll,
-                           double init_pitch,
-                           double init_yaw,
-                           const std::vector<double> &init_qpos):
-        activate_shield_(activate_shield),
-        sample_time_(sample_time),
-        path_s_(0),
-        path_s_discrete_(0)
-{
-    ///////////// Build robot reach
-    YAML::Node robot_config = YAML::LoadFile(robot_config_file);
-    std::string robot_name = robot_config["robot_name"].as<std::string>();
-    nb_joints_ = robot_config["nb_joints"].as<int>();
-    std::vector<double> transformation_matrices = robot_config["transformation_matrices"].as<std::vector<double>>();
-    std::vector<double> transformation_matrices_joints = robot_config["transformation_matrices_joints"].as<std::vector<double>>();
-    std::vector<double> enclosures = robot_config["enclosures"].as<std::vector<double>>();
-    double secure_radius = robot_config["secure_radius"].as<double>();
-    robot_reach_ = new RobotReach(transformation_matrices,
-                                  transformation_matrices_joints,
-                                  nb_joints_,
-                                  enclosures,
-                                  init_x, init_y, init_z,
-                                  init_roll, init_pitch, init_yaw,
-                                  secure_radius);
-    ////////////// Setting trajectory variables
-    YAML::Node trajectory_config = YAML::LoadFile(trajectory_config_file);
-    max_s_stop_ = trajectory_config["max_s_stop"].as<double>();
-    q_min_allowed_ = trajectory_config["q_min_allowed"].as<std::vector<double>>();
-    q_max_allowed_ = trajectory_config["q_max_allowed"].as<std::vector<double>>();
-    v_max_allowed_ = trajectory_config["v_max_allowed"].as<std::vector<double>>();
-    a_max_allowed_ = trajectory_config["a_max_allowed"].as<std::vector<double>>();
-    j_max_allowed_ = trajectory_config["j_max_allowed"].as<std::vector<double>>();
-    a_max_ltt_ = trajectory_config["a_max_ltt"].as<std::vector<double>>();
-    j_max_ltt_ = trajectory_config["j_max_ltt"].as<std::vector<double>>();
-    ltp_ = long_term_planner::LongTermPlanner(nb_joints_, sample_time, q_min_allowed_, q_max_allowed_, v_max_allowed_, a_max_ltt_, j_max_ltt_);
-    v_iso_ = trajectory_config["v_iso"].as<double>();
-    safety_method_ = static_cast<Safety_method>(trajectory_config["safety_method"].as<int>());
-    RobotReach::Velocity_method velocity_method = static_cast<RobotReach::Velocity_method>(trajectory_config["velocity_method"].as<int>());
-    robot_reach_->setVelocityMethod(velocity_method);
-    spdlog::info("safety_method = {}", safety_method_);
-    spdlog::info("velocity_method = {}", velocity_method);
-    // Initialize the long term trajectory
-    std::vector<Motion> long_term_traj;
-    long_term_traj.push_back(Motion(0.0, init_qpos));
-    if(safety_method_ == STANDARD) {
-        long_term_trajectory_ = LongTermTraj(long_term_traj, sample_time_);
-    } else {
-        long_term_trajectory_ = LongTermTraj(long_term_traj, sample_time_, *robot_reach_);
-    }
-    //////////// Build human reach
-    YAML::Node human_config = YAML::LoadFile(mocap_config_file);
-    double measurement_error_pos = human_config["measurement_error_pos"].as<double>();
-    double measurement_error_vel = human_config["measurement_error_vel"].as<double>();
-    double delay = human_config["delay"].as<double>();
-
-    std::vector<std::string> joint_name_vec = human_config["joint_names"].as<std::vector<std::string>>();
-    std::map<std::string, int> joint_names;
-    for(std::size_t i = 0; i < joint_name_vec.size(); ++i) {
-        joint_names[joint_name_vec[i]] = i;
-    }
-
-    std::vector<double> joint_v_max = human_config["joint_v_max"].as<std::vector<double>>();
-    std::vector<double> joint_a_max = human_config["joint_a_max"].as<std::vector<double>>();
-    // Build bodies
-    const YAML::Node& bodies = human_config["bodies"];
-    std::map<std::string, reach_lib::jointPair> body_link_joints;
-    std::map<std::string, double> thickness;
-    for (YAML::const_iterator it = bodies.begin(); it != bodies.end(); ++it) {
-        const YAML::Node& body = *it;
-        body_link_joints[body["name"].as<std::string>()] = reach_lib::jointPair(joint_names[body["proximal"].as<std::string>()], joint_names[body["distal"].as<std::string>()]);
-        thickness[body["name"].as<std::string>()] = body["thickness"].as<double>();
-    }
-    // Build extremities
-    const YAML::Node& extremities = human_config["extremities"];
-    std::vector<std::string> extremity_base_names;
-    std::vector<std::string> extremity_end_names;
-    std::vector<double> extremity_length;
-    for (YAML::const_iterator it = extremities.begin(); it != extremities.end(); ++it) {
-        const YAML::Node& extremity = *it;
-        extremity_base_names.push_back(extremity["base"].as<std::string>());
-        extremity_end_names.push_back(extremity["end"].as<std::string>());
-        extremity_length.push_back(extremity["length"].as<double>());
-    }
-    human_reach_ = new HumanReach(joint_names.size(),
-                                  body_link_joints,
-                                  thickness,
-                                  joint_v_max,
-                                  joint_a_max,
-                                  extremity_base_names,
-                                  extremity_end_names,
-                                  extremity_length,
-                                  measurement_error_pos,
-                                  measurement_error_vel,
-                                  delay);
-    ///////////// Build verifier
-    verify_ = new safety_shield::VerifyISO();
-    /////////// Other settings
-    sliding_window_k_ = (int) std::floor(max_s_stop_/sample_time_);
-    std::vector<double> prev_dq;
-
-    for(int i = 0; i < 6; i++) {
-        prev_dq.push_back(0.0);
-        alpha_i_.push_back(1.0);
-    }
-    alpha_i_.push_back(1.0);
-    is_safe_ = !activate_shield_;
-    is_static_safe_ = is_safe_;
-    is_PFL_safe_ = is_safe_;
-    if(safety_method_ == STANDARD) {
-        computesPotentialTrajectory(is_safe_, prev_dq);
-        next_motion_ = determineNextMotion(is_safe_);
-    } else {
-        computesPotentialTrajectoryForPFL(is_static_safe_, is_PFL_safe_, prev_dq, nullptr, nullptr);
-        next_motion_ = determineNextMotionForPFL(is_static_safe_, is_PFL_safe_);
-    }
-    spdlog::info("Safety shield created.");
-}
-
-void SafetyShield::reset(bool activate_shield,
-                         double init_x,
-                         double init_y,
-                         double init_z,
-                         double init_roll,
-                         double init_pitch,
-                         double init_yaw,
-                         const std::vector<double> &init_qpos,
-                         double current_time) {
-    robot_reach_->reset(init_x, init_y, init_z, init_roll, init_pitch, init_yaw);
-    human_reach_->reset();
-    std::vector<double> prev_dq;
-    for(int i = 0; i < 6; i++) {
-        prev_dq.push_back(0.0);
-        alpha_i_.push_back(1.0);
-    }
-    alpha_i_.push_back(1.0);
-    is_safe_ = !activate_shield_;
-    is_static_safe_ = is_safe_;
-    is_PFL_safe_ = is_safe_;
-    new_ltt_ = false;
-    new_goal_ = false;
-    new_ltt_processed_ = false;
-    recovery_path_correct_ = false;
-    path_s_ = 0;
-    path_s_discrete_ = 0;
-    cycle_begin_time_ = current_time;
-    recovery_path_ = Path();
-    failsafe_path_ = Path();
-    failsafe_path_2_ = Path();
-    safe_path_ = Path();
-    failsafe_path_static_ = Path();
-    failsafe_path_pfl_ = Path();
-    failsafe_path_2_static_ = Path();
-    failsafe_path_2_pfl_ = Path();
-    // Initialize the long term trajectory
-    std::vector<Motion> long_term_traj;
-    long_term_traj.push_back(Motion(0.0, init_qpos));
-    if(safety_method_ == STANDARD) {
-        long_term_trajectory_ = LongTermTraj(long_term_traj, sample_time_);
-    } else {
-        long_term_trajectory_ = LongTermTraj(long_term_traj, sample_time_, *robot_reach_);
-    }
-    if(safety_method_ == STANDARD) {
-        computesPotentialTrajectory(is_safe_, prev_dq);
-        next_motion_ = determineNextMotion(is_safe_);
-    } else {
-        computesPotentialTrajectoryForPFL(is_static_safe_, is_PFL_safe_, prev_dq, nullptr, nullptr);
-        next_motion_ = determineNextMotionForPFL(is_static_safe_, is_PFL_safe_);
-    }
-    spdlog::info("Safety shield resetted.");
-}
-
-bool SafetyShield::planSafetyShield(double pos, double vel, double acc, double ve, double a_max, double j_max, Path &path) {
-    if (a_max < 0 || fabs(acc) > a_max) {
-        return false;
-    }
-    try {
-        path.setCurrent(false);
-        path.setPosition(pos);
-        path.setVelocity(vel);
-        path.setAcceleration(acc);
-        double epsilon = 1e-6;
-
-        // If current velocity is close to goal and acceleration is approx zero -> Do nothing
-        if (fabs(vel - ve) < epsilon && fabs(acc) < epsilon) {
-            std::array<double,6> new_phases = { sample_time_, sample_time_, sample_time_, 0.0, 0.0, 0.0};
-            path.setPhases(new_phases);
-            return true;
-        }
-        // Time to reach zero acceleration with maximum jerk
-        double t_to_a_0 = roundToTimestep(abs(acc)/j_max);
-        // Velocity at zero acceleration
-        double v_at_a_0 = vel + acc*t_to_a_0/2;
-
-        // If the velocity at zero acceleration is the goal velocity, use t_to_a_0 and max jerk and return
-        if (fabs(v_at_a_0 - ve) < epsilon){
-            std::array<double,6> new_phases = { t_to_a_0, t_to_a_0, t_to_a_0, -acc/t_to_a_0, 0, 0 };
-            path.setPhases(new_phases);
-            return true;
-        }
-
-        double a_peak,t01,t12,t23,v01,v12,v23,correct_int;
-        // If we are accelerating
-        if (acc > 0) {
-            // Reducing the acceleration to zero with max jerk would lead to a velocity larger than target
-            if (v_at_a_0 > ve + epsilon) {
-                if (vel - (a_max*a_max + acc*acc/2)/j_max > ve) {
-                    a_peak = -a_max;
-                    t01 = roundToTimestep((acc - a_peak)/j_max);
-                    t23 = roundToTimestep(-a_peak/j_max);
-                    v01 = (acc + a_peak)*t01/2;
-                    v23 = a_peak*t23/2;
-                    t12 = roundToTimestep((ve - vel - v01 - v23)/a_peak);
-                    v12 = a_peak*t12;
-                }
-                else {
-                    a_peak = -sqrt(fabs((vel - ve)*j_max + acc*acc/2));
-                    t01 = roundToTimestep((acc - a_peak)/j_max);
-                    t23 = roundToTimestep(-a_peak/j_max);
-                    v01 = (acc + a_peak)*t01/2;
-                    v23 = a_peak*t23/2;
-                    t12 = 0;
-                    v12 = 0;
-                }
-            }
-            else {
-                if (vel + (a_max*a_max - acc*acc/2)/j_max < ve) {
-                    a_peak = a_max;
-                    t01 = roundToTimestep((a_peak - acc)/j_max);
-                    t23 = roundToTimestep(a_peak/j_max);
-                    v01 = (acc + a_peak)*t01/2;
-                    v23 = a_peak*t23/2;
-                    t12 = roundToTimestep((ve - vel - v01 - v23)/a_peak);
-                    v12 = a_peak*t12;
-                }
-                else {
-                    a_peak = sqrt(fabs((ve - vel)*j_max + acc*acc/2));
-                    t01 = roundToTimestep((a_peak - acc)/j_max);
-                    t23 = roundToTimestep(a_peak/j_max);
-                    v01 = (acc + a_peak)*t01/2;
-                    v23 = a_peak*t23/2;
-                    t12 = 0;
-                    v12 = 0;
-                }
-            }
-        }
-            // If we are decelerating (a <= 0)
-        else {
-            if (v_at_a_0 > ve + epsilon) {
-                if (vel - (a_max*a_max - acc*acc/2)/j_max > ve) {
-                    a_peak = -a_max;
-                    t01 = roundToTimestep((acc - a_peak)/j_max);
-                    t23 = roundToTimestep(-a_peak/j_max);
-                    v01 = (acc + a_peak)*t01/2;
-                    v23 = a_peak*t23/2;
-                    t12 = roundToTimestep((ve - vel - v01 - v23)/(a_peak));
-                    v12 = a_peak*t12;
-                }
-                else {
-                    a_peak = -sqrt(fabs((vel - ve)*j_max + acc*acc/2));
-                    t01 = roundToTimestep((acc - a_peak)/j_max);
-                    t23 = roundToTimestep(-a_peak/j_max);
-                    v01 = (acc + a_peak)*t01/2;
-                    v23 = a_peak*t23/2;
-                    t12 = 0;
-                    v12 = 0;
-                }
-            }
-            else {
-                if (vel + (a_max*a_max + acc*acc/2)/j_max < ve) {
-                    a_peak = a_max;
-                    t01 = roundToTimestep((a_peak - acc)/j_max);
-                    t23 = roundToTimestep(a_peak/j_max);
-                    v01 = (acc + a_peak)*t01/2;
-                    v23 = a_peak*t23/2;
-                    t12 = roundToTimestep((ve - vel - v01 - v23)/a_peak);
-                    v12 = a_peak*t12;
-                }
-                else {
-                    a_peak = sqrt(fabs((ve - vel)*j_max + acc*acc/2));
-                    t01 = roundToTimestep((a_peak - acc)/j_max);
-                    t23 = roundToTimestep(a_peak/j_max);
-                    v01 = (acc + a_peak)*t01/2;
-                    v23 = a_peak*t23/2;
-                    t12 = 0;
-                    v12 = 0;
-                }
-            }
-        }
-        correct_int = ve - vel - v01 - v12 - v23;
-        a_peak += correct_int/((t01 + t23)/2 + t12 + epsilon);
-        double j01;
-        if (t01 >= epsilon) {
-            j01 = (a_peak - acc)/(t01);
-        } else {
-            j01 = 0;
-        }
-        double j12 = 0;
-        double j23;
-        if (t23 >= epsilon) {
-            j23 = -a_peak/(t23);
-        } else {
-            j23 = 0;
-        }
-        if (t01 < 0 || t12 < 0 || t23 < 0) {
-            spdlog::debug("planSafetyShield calculated time negative. t01 = {}, t12 = {}, t23 = {}", t01, t12, t23);
-            return false;
-        }
-        std::array<double,6> new_phases = { t01, t01+t12, t01+t12+t23, j01, j12, j23 };
-        path.setPhases(new_phases);
-        return true;
-    } catch (const std::exception &exc) {
-        spdlog::error("Exception in SafetyShield::planSafetyShield: {}", exc.what());
-        return false;
-    }
-}
-
-void SafetyShield::calculateMaxAccJerk(const std::vector<double> &prev_speed, const std::vector<double>& a_max_part, const std::vector<double>& j_max_part, double& a_max_manoeuvre, double& j_max_manoeuvre) {
-    double new_c, new_d;
-    // Calculate dds_max
-    double denom = std::abs(prev_speed[0]) + std::abs(a_max_part[0]) * max_s_stop_ + 1E-9;
-    double min_c = (a_max_allowed_[0] - std::abs(a_max_part[0])) / denom;
-    for (int i = 1; i < a_max_allowed_.size(); i++) {
-        denom = std::abs(prev_speed[i]) + std::abs(a_max_part[i]) * max_s_stop_;
-        new_c = (a_max_allowed_[i] - std::abs(a_max_part[i])) / denom;
-        min_c = (new_c < min_c) ? new_c : min_c;
-    }
-    a_max_manoeuvre = (min_c < 0) ? 0 : min_c;
-    // Calculate ddds_max
-    denom = std::abs(prev_speed[0]) + std::abs(a_max_part[0]) * max_s_stop_ + 1E-9;
-    double min_d = (j_max_allowed_[0] - 3*std::abs(a_max_part[0])*a_max_manoeuvre - std::abs(j_max_part[0])) / denom;
-    for (int i = 1; i < a_max_allowed_.size(); i++) {
-        denom = std::abs(prev_speed[i]) + std::abs(a_max_part[i]) * max_s_stop_;
-        new_d = (j_max_allowed_[i] - 3*std::abs(a_max_part[i])*a_max_manoeuvre - std::abs(j_max_part[i])) / denom;
-        min_d = (new_d < min_d) ? new_d : min_d;
-    }
-    j_max_manoeuvre = (min_d < 0) ? 0 : min_d;
-}
-
-Motion SafetyShield::computesPotentialTrajectory(bool v, const std::vector<double> &prev_speed)
-{
-    try {
-        // s_int indicates the index of the entire traveled way
-        while (path_s_ >= (path_s_discrete_+1)*sample_time_) {
-            long_term_trajectory_.increasePosition();
-            if (new_ltt_) {
-                new_long_term_trajectory_.increasePosition();
-            }
-            path_s_discrete_++;
-        }
-        //If verified safe, take the recovery path, otherwise, take the failsafe path
-        if (v && recovery_path_correct_) {
-            recovery_path_.setCurrent(true);
-            //discard old FailsafePath and replace with new one
-            failsafe_path_ = failsafe_path_2_;
-            //repair path already incremented
-        }
-        else {
-            failsafe_path_.setCurrent(true);
-            //discard RepairPath
-            recovery_path_.setCurrent(false);
-            failsafe_path_.increment(sample_time_);
-        }
-
-        //find maximum acceleration and jerk authorised
-        double a_max_manoeuvre, j_max_manoeuvre;
-        // One could use new_long_term_trajectory_.getMaxAccelerationWindow(path_s_discrete_) instead of a_max_ltt but there are many bugs to be solved before.
-        if (!new_ltt_) {
-            calculateMaxAccJerk(prev_speed, a_max_ltt_, j_max_ltt_, a_max_manoeuvre, j_max_manoeuvre);
-        } else {
-            calculateMaxAccJerk(prev_speed, a_max_ltt_, j_max_ltt_, a_max_manoeuvre, j_max_manoeuvre);
-        }
-
-        //Desired movement, one timestep
-        // if not already on the repair path, plan a repair path
-        if (!recovery_path_.isCurrent()) {
-            // plan repair path and replace
-            recovery_path_correct_ = planSafetyShield(failsafe_path_.getPosition(), failsafe_path_.getVelocity(),
-                                                      failsafe_path_.getAcceleration(), 1, a_max_manoeuvre, j_max_manoeuvre, recovery_path_);
-        }
-        // Only plan new failsafe trajectory if the recovery path planning was successful.
-        if (recovery_path_correct_) {
-            //advance one step on repair path
-            recovery_path_.increment(sample_time_);
-
-            // plan new failsafe path for STP
-            bool failsafe_2_planning_success = planSafetyShield(recovery_path_.getPosition(), recovery_path_.getVelocity(), recovery_path_.getAcceleration(), 0, a_max_manoeuvre, j_max_manoeuvre, failsafe_path_2_);
-            // Check the validity of the planned path
-            if (!failsafe_2_planning_success || recovery_path_.getPosition() < failsafe_path_.getPosition()) {
-                recovery_path_correct_ = false;
-            }
-        }
-        // If all planning was correct, use new failsafe path with single recovery step
-        if (recovery_path_correct_) {
-            potential_path_ = failsafe_path_2_;
-        }
-        else {
-            // If planning failed, use previous failsafe path
-            potential_path_ = failsafe_path_;
-        }
-
-        //// Calculate start and goal pos of intended motion
-        // Fill potential buffer with position and velocity from last failsafe path. This value is not really used.
-        double final_s_d, final_ds_d, final_dds_d;
-        double ddds_d = failsafe_path_.getJerk();
-        // Calculate goal
-        potential_path_.getFinalMotion(final_s_d, final_ds_d, final_dds_d);
-        Motion goal_motion;
-        if (new_ltt_) {
-            goal_motion = new_long_term_trajectory_.interpolate(final_s_d, final_ds_d, final_dds_d, ddds_d, v_max_allowed_, a_max_allowed_, j_max_allowed_);
-        } else {
-            goal_motion = long_term_trajectory_.interpolate(final_s_d, final_ds_d, final_dds_d, ddds_d, v_max_allowed_, a_max_allowed_, j_max_allowed_);
-        }
-        goal_motion.setTime(potential_path_.getPhase(3));
-        return goal_motion;
-    } catch (const std::exception &exc) {
-        spdlog::error("Exception in SafetyShield::computesPotentialTrajectory: {}", exc.what());
-        throw exc;
-    }
-}
-
-void SafetyShield::computesPotentialTrajectoryForPFL(bool v_static, bool v_pfl, const std::vector<double>& prev_speed, Motion* goal_motion_static, Motion* goal_motion_pfl) {
-    // s_int indicates the index of the entire traveled way
-    while (path_s_ >= (path_s_discrete_+1)*sample_time_) {
-        long_term_trajectory_.increasePosition();
-        if (new_ltt_) {
-            new_long_term_trajectory_.increasePosition();
-        }
-        path_s_discrete_++;
-    }
-    //If verified safe, take the recovery path, otherwise, take the failsafe path
-    // TODO: zweite Unterscheidung, entweder failsafe_path static oder PFL? --> failsafe_path_2_static_ und failsafe_path_2_pfl_
-    if (v_static && v_pfl && recovery_path_correct_) {
-        recovery_path_.setCurrent(true);
-        //discard old FailsafePath and replace with new one
-        failsafe_path_static_ = failsafe_path_2_static_;
-        failsafe_path_pfl_ = failsafe_path_2_pfl_;
-        //repair path already incremented
-    }
-        // if v_static but not v_pfl, we can take the failsafe-path of the PFL criterion
-    else if(v_static && !v_pfl) {
-        failsafe_path_pfl_.setCurrent(true);
-        failsafe_path_static_.setCurrent(false);
-        //discard RepairPath
-        recovery_path_.setCurrent(false);
-        failsafe_path_pfl_.increment(sample_time_);
-        failsafe_path_static_.increment(sample_time_);
-    }
-        // if no criterion is safe, we take the failsafe-path of the static criterion
-    else {
-        failsafe_path_pfl_.setCurrent(false);
-        failsafe_path_static_.setCurrent(true);
-        //discard RepairPath
-        recovery_path_.setCurrent(false);
-        failsafe_path_pfl_.increment(sample_time_);
-        failsafe_path_static_.increment(sample_time_);
-    }
-
-    //find maximum acceleration and jerk authorised
-    double a_max_manoeuvre, j_max_manoeuvre;
-    // One could use new_long_term_trajectory_.getMaxAccelerationWindow(path_s_discrete_) instead of a_max_ltt but there are many bugs to be solved before.
-    if (!new_ltt_) {
-        calculateMaxAccJerk(prev_speed, a_max_ltt_, j_max_ltt_, a_max_manoeuvre, j_max_manoeuvre);
-    } else {
-        calculateMaxAccJerk(prev_speed, a_max_ltt_, j_max_ltt_, a_max_manoeuvre, j_max_manoeuvre);
-    }
-
-    //Desired movement, one timestep
-    // if not already on the repair path, plan a repair path
-    if (!recovery_path_.isCurrent()) {
-        double pos, vel, acc;
-        if(failsafe_path_pfl_.isCurrent()) {
-            pos = failsafe_path_pfl_.getPosition();
-            vel = failsafe_path_pfl_.getVelocity();
-            acc = failsafe_path_pfl_.getAcceleration();
-        } else if(failsafe_path_static_.isCurrent()) {
-            pos = failsafe_path_static_.getPosition();
-            vel = failsafe_path_static_.getVelocity();
-            acc = failsafe_path_static_.getAcceleration();
-        } else {
-            spdlog::error("Error in computesPotentialTrajectoryForPFL: no failsafe-path is set when planning repair path?");
-        }
-
-        // plan repair path and replace
-        recovery_path_correct_ = planSafetyShield(pos, vel, acc, 1, a_max_manoeuvre, j_max_manoeuvre, recovery_path_);
-    }
-    // Only plan new failsafe trajectory if the recovery path planning was successful.
-    if (recovery_path_correct_) {
-        //advance one step on repair path
-        recovery_path_.increment(sample_time_);
-
-        double s_d, ds_d, dds_d;
-        // Calculate goal
-        // if v_max is under v_iso, velocity_criteria is true and we can skip calculating getMotionUnderVel()
-        is_under_iso_velocity_ = false;
-        // safety_method == LTT_MAXIMUM or STP_MAXIMUM
-        // v_max is maximum of LTT or STP and vel_s_dot is how much path velocity needs to be scaled to be under v_iso
-        failsafe_path_2_pfl_.getFinalMotion(s_d, ds_d, dds_d);
-        double v_max;
-        if(safety_method_ == LTT_MAXIMUM && !new_ltt_) {
-            v_max = long_term_trajectory_.getMaxofMaximumCartesianVelocity();
-        } else if(safety_method_ == LTT_MAXIMUM && new_ltt_) {
-            v_max = new_long_term_trajectory_.getMaxofMaximumCartesianVelocity();
-        } else if(safety_method_ == STP_MAXIMUM && !new_ltt_) {
-            v_max = long_term_trajectory_.getMaxofMaximumCartesianVelocityWithS(s_d);
-        } else {
-            // safety_method_ == STP_MAXIMUM && new_ltt_
-            v_max = new_long_term_trajectory_.getMaxofMaximumCartesianVelocityWithS(s_d);
-        }
-        is_under_iso_velocity_ = v_max <= v_iso_;
-        double v_limit;
-        bool failsafe_2_pfl_planning_success = true;
-        // TODO: wie mach ich das mit is_under_iso_velocity, welcher Code darf ich skippen?
-        if(!is_under_iso_velocity_) {
-            v_limit = v_iso_ / v_max;
-            failsafe_2_pfl_planning_success = planSafetyShield(recovery_path_.getPosition(), recovery_path_.getVelocity(), recovery_path_.getAcceleration(), v_limit, a_max_manoeuvre, j_max_manoeuvre, failsafe_path_2_pfl_);
-        }
-
-        // plan new failsafe path for STP
-        bool failsafe_2_static_planning_success = planSafetyShield(recovery_path_.getPosition(), recovery_path_.getVelocity(), recovery_path_.getAcceleration(), 0, a_max_manoeuvre, j_max_manoeuvre, failsafe_path_2_static_);
-
-        // Check the validity of the planned path
-        if (!failsafe_2_static_planning_success || !failsafe_2_pfl_planning_success || recovery_path_.getPosition() < failsafe_path_static_.getPosition() || recovery_path_.getPosition() < failsafe_path_pfl_.getPosition()) {
-            recovery_path_correct_ = false;
-        }
-    }
-    // If all planning was correct, use new failsafe path with single recovery step
-    if (recovery_path_correct_) {
-        potential_path_static_ = failsafe_path_2_static_;
-        potential_path_pfl_ = failsafe_path_2_pfl_;
-    }
-    else {
-        // TODO: weitere Fallunterscheidung: je nachdem welches planning failed?
-        // If planning failed, use previous failsafe path
-        potential_path_static_ = failsafe_path_static_;
-        potential_path_pfl_ = failsafe_path_pfl_;
-    }
-
-    if(goal_motion_static == nullptr) {
-        return;
-    }
-
-    //// Calculate start and goal pos of intended motion
-    // Fill potential buffer with position and velocity from last failsafe path. This value is not really used.
-    double final_static_s_d, final_static_ds_d, final_static_dds_d;
-    double final_pfl_s_d, final_pfl_ds_d, final_pfl_dds_d;
-    double ddds_d;
-    // TODO: neither could be set because its old failsafe-path?
-    ddds_d = failsafe_path_static_.getJerk();
-    /*
-    if(failsafe_path_pfl_.isCurrent()) {
-        ddds_d = failsafe_path_pfl_.getJerk();
-    } else if(failsafe_path_static_.isCurrent()) {
-        ddds_d = failsafe_path_static_.getJerk();
-    } else {
-        spdlog::error("Error in computesPotentialTrajectoryForPFL: no failsafe-path is set when calculating motions?");
-    }
-     */
-    potential_path_static_.getFinalMotion(final_static_s_d, final_static_ds_d, final_static_dds_d);
-    potential_path_pfl_.getFinalMotion(final_pfl_s_d, final_pfl_ds_d, final_pfl_dds_d);
-
-    if (new_ltt_) {
-        *goal_motion_static = new_long_term_trajectory_.interpolate(final_static_s_d, final_static_ds_d, final_static_dds_d, ddds_d, v_max_allowed_, a_max_allowed_, j_max_allowed_);
-        if(!is_under_iso_velocity_) {
-            *goal_motion_pfl = new_long_term_trajectory_.interpolate(final_pfl_s_d, final_pfl_ds_d, final_pfl_dds_d, ddds_d, v_max_allowed_, a_max_allowed_, j_max_allowed_);
-        }
-    } else {
-        *goal_motion_static = long_term_trajectory_.interpolate(final_static_s_d, final_static_ds_d, final_static_dds_d, ddds_d, v_max_allowed_, a_max_allowed_, j_max_allowed_);
-        if(!is_under_iso_velocity_) {
-            *goal_motion_pfl = long_term_trajectory_.interpolate(final_pfl_s_d, final_pfl_ds_d, final_pfl_dds_d, ddds_d, v_max_allowed_, a_max_allowed_, j_max_allowed_);
-        }
-    }
-    goal_motion_static->setTime(potential_path_static_.getPhase(3));
-    if(!is_under_iso_velocity_) {
-        goal_motion_pfl->setTime(potential_path_pfl_.getPhase(3));
-    }
-}
-
-bool SafetyShield::checkMotionForJointLimits(Motion& motion) {
-    for (int i = 0; i < nb_joints_; i++) {
-        if (motion.getAngle()[i] < q_min_allowed_[i] || motion.getAngle()[i] > q_max_allowed_[i]) {
-            return false;
-        }
-    }
-    return true;
-}
-
-
-Motion SafetyShield::determineNextMotion(bool is_safe) {
-    Motion next_motion;
-    double s_d, ds_d, dds_d, ddds_d;
-    if (is_safe) {
-        // Fill potential buffer with position and velocity from recovery path
-        if (recovery_path_.getPosition() >= failsafe_path_.getPosition()) {
-            s_d = recovery_path_.getPosition();
-            ds_d = recovery_path_.getVelocity();
-            dds_d = recovery_path_.getAcceleration();
-            ddds_d = recovery_path_.getJerk();
-        }
-        else {
-            potential_path_.increment(sample_time_);
-            s_d = potential_path_.getPosition();
-            ds_d = potential_path_.getVelocity();
-            dds_d = potential_path_.getAcceleration();
-            ddds_d = potential_path_.getJerk();
-        }
-
-        // Interpolate from new long term buffer
-        if (new_ltt_) {
-            next_motion = new_long_term_trajectory_.interpolate(s_d, ds_d, dds_d, ddds_d, v_max_allowed_, a_max_allowed_, j_max_allowed_);
-        } else {
-            next_motion = long_term_trajectory_.interpolate(s_d, ds_d, dds_d, ddds_d, v_max_allowed_, a_max_allowed_, j_max_allowed_);
-        }
-        // Set potential path as new verified safe path
-        safe_path_ = potential_path_;
-    } else {
-        // interpolate from old safe path
-        safe_path_.increment(sample_time_);
-        s_d = safe_path_.getPosition();
-        ds_d = safe_path_.getVelocity();
-        dds_d = safe_path_.getAcceleration();
-        ddds_d = safe_path_.getJerk();
-        next_motion = long_term_trajectory_.interpolate(s_d, ds_d, dds_d, ddds_d, v_max_allowed_, a_max_allowed_, j_max_allowed_);
-    }
-    /// !!! Set s to the new path position !!!
-    path_s_ = s_d;
-    // Return the calculated next motion
-    return next_motion;
-}
-
-Motion SafetyShield::determineNextMotionForPFL(bool is_safe_static, bool is_safe_pfl) {
-    Motion next_motion;
-    double s_d, ds_d, dds_d, ddds_d;
-    // TODO: depending on verification, set safe_path_ to potential_path_static or potential_path_pfl?
-    // TODO: macht diese Unterteilung Sinn?
-    double pos;
-    // TODO: neither could be set because its old failsafe-path?
-    pos = failsafe_path_static_.getPosition();
-    /*
-    if(failsafe_path_pfl_.isCurrent()) {
-        pos = failsafe_path_pfl_.getPosition();
-    } else if(failsafe_path_static_.isCurrent()) {
-        pos = failsafe_path_static_.getPosition();
-    } else {
-        spdlog::error("Error in determineNextMotionForPFL: no failsafe-path is set?");
-    }
-     */
-    if (is_safe_static && is_safe_pfl) {
-        // Fill potential buffer with position and velocity from recovery path
-        if (recovery_path_.getPosition() >= pos) {
-            s_d = recovery_path_.getPosition();
-            ds_d = recovery_path_.getVelocity();
-            dds_d = recovery_path_.getAcceleration();
-            ddds_d = recovery_path_.getJerk();
-        } else {
-            potential_path_pfl_.increment(sample_time_);
-            potential_path_static_.increment(sample_time_);
-            s_d = potential_path_pfl_.getPosition();
-            ds_d = potential_path_pfl_.getVelocity();
-            dds_d = potential_path_pfl_.getAcceleration();
-            ddds_d = potential_path_pfl_.getJerk();
-        }
-
-        // Interpolate from new long term buffer
-        if (new_ltt_) {
-            next_motion = new_long_term_trajectory_.interpolate(s_d, ds_d, dds_d, ddds_d, v_max_allowed_,
-                                                                a_max_allowed_, j_max_allowed_);
-        } else {
-            next_motion = long_term_trajectory_.interpolate(s_d, ds_d, dds_d, ddds_d, v_max_allowed_, a_max_allowed_,
-                                                            j_max_allowed_);
-        }
-        // Set potential path as new verified safe path
-        safe_path_ = potential_path_pfl_;
-
-    } else if(is_safe_static && !is_safe_pfl) {
-        // Fill potential buffer with position and velocity from recovery path
-        if (recovery_path_.getPosition() >= pos) {
-            s_d = recovery_path_.getPosition();
-            ds_d = recovery_path_.getVelocity();
-            dds_d = recovery_path_.getAcceleration();
-            ddds_d = recovery_path_.getJerk();
-        } else {
-            potential_path_pfl_.increment(sample_time_);
-            potential_path_static_.increment(sample_time_);
-            s_d = potential_path_static_.getPosition();
-            ds_d = potential_path_static_.getVelocity();
-            dds_d = potential_path_static_.getAcceleration();
-            ddds_d = potential_path_static_.getJerk();
-        }
-
-        // Interpolate from new long term buffer
-        if (new_ltt_) {
-            next_motion = new_long_term_trajectory_.interpolate(s_d, ds_d, dds_d, ddds_d, v_max_allowed_,
-                                                                a_max_allowed_, j_max_allowed_);
-        } else {
-            next_motion = long_term_trajectory_.interpolate(s_d, ds_d, dds_d, ddds_d, v_max_allowed_, a_max_allowed_,
-                                                            j_max_allowed_);
-        }
-        // Set potential path as new verified safe path
-        safe_path_ = potential_path_static_;
-    } else {
-        // interpolate from old safe path
-        safe_path_.increment(sample_time_);
-        s_d = safe_path_.getPosition();
-        ds_d = safe_path_.getVelocity();
-        dds_d = safe_path_.getAcceleration();
-        ddds_d = safe_path_.getJerk();
-        next_motion = long_term_trajectory_.interpolate(s_d, ds_d, dds_d, ddds_d, v_max_allowed_, a_max_allowed_, j_max_allowed_);
-    }
-    /// !!! Set s to the new path position !!!
-    path_s_ = s_d;
-    // Return the calculated next motion
-    return next_motion;
-}
-
-Motion SafetyShield::step(double cycle_begin_time) {
-    if(safety_method_ == STANDARD) {
-        return standardStep(cycle_begin_time);
-    } else {
-        return PFLstep(cycle_begin_time);
-    }
-}
-
-Motion SafetyShield::standardStep(double cycle_begin_time) {
-    cycle_begin_time_ = cycle_begin_time;
-    try {
-        // Get current motion
-        Motion current_motion = getCurrentMotion();
-        // If the new LTT was processed at least once and is labeled safe, replace old LTT with new one.
-        if (new_ltt_ && new_ltt_processed_) {
-            if (is_safe_ || current_motion.isStopped()) {
-                long_term_trajectory_ = new_long_term_trajectory_;
-                new_ltt_ = false;
-                new_goal_ = false;
-            }
-        }
-        // Check if there is a new goal motion
-        if (new_goal_) {
-            // Check if current motion has acceleration and jerk values that lie in the plannable ranges
-            bool is_plannable = checkCurrentMotionForReplanning(current_motion);
-            if (is_plannable) {
-                // Check if the starting position of the last replanning was very close to the current position
-                bool last_close = true;
-                if (new_ltt_ == true) {
-                    for (int i = 0; i < current_motion.getAngle().size(); i++) {
-                        if (std::abs(current_motion.getAngle()[i] - last_replan_start_motion_.getAngle()[i]) > 0.01) {
-                            last_close = false;
-                            break;
-                        }
-                    }
-                } else {
-                    last_close = false;
-                }
-                // Only replan if the current joint position is different from the last.
-                bool success = true;
-                if (!last_close) {
-                    success = calculateLongTermTrajectory(current_motion.getAngle(), current_motion.getVelocity(),
-                                                          current_motion.getAcceleration(), new_goal_motion_.getAngle(), new_long_term_trajectory_);
-                    if (success) {
-                        last_replan_start_motion_ = current_motion;
-                    }
-                }
-                if (success) {
-                    new_ltt_ = true;
-                    new_ltt_processed_ = false;
-                } else {
-                    new_ltt_ = false;
-                }
-            } else {
-                new_ltt_ = false;
-            }
-        }
-        // If there is a new long term trajectory (LTT), always override is_safe with false.
-        if (new_ltt_ && !new_ltt_processed_) {
-            is_safe_ = false;
-        }
-        // Compute a new potential trajectory
-        Motion goal_motion = computesPotentialTrajectory(is_safe_, next_motion_.getVelocity());
-        if (activate_shield_) {
-            // Check motion for joint limits
-            if (!checkMotionForJointLimits(goal_motion)) {
-                is_safe_ = false;
-            } else {
-                // Compute the robot reachable set for the potential trajectory
-                robot_capsules_ = robot_reach_->reach(current_motion, goal_motion, (goal_motion.getS()-current_motion.getS()), alpha_i_);
-                // Compute the human reachable sets for the potential trajectory
-                // humanReachabilityAnalysis(t_command, t_brake)
-                human_reach_->humanReachabilityAnalysis(cycle_begin_time_, goal_motion.getTime());
-                human_capsules_ = human_reach_->getAllCapsules();
-                // Verify if the robot and human reachable sets are collision free
-                is_safe_ = verify_->verify_human_reach(robot_capsules_, human_capsules_);
-            }
-        } else {
-            is_safe_ = true;
-        }
-        // Select the next motion based on the verified safety
-        next_motion_ = determineNextMotion(is_safe_);
-        next_motion_.setTime(cycle_begin_time);
-        new_ltt_processed_ = true;
-        return next_motion_;
-    } catch (const std::exception &exc) {
-        spdlog::error("Exception in SafetyShield::getNextCycleSTANDARD: {}", exc.what());
-    }
-}
-
-Motion SafetyShield::PFLstep(double cycle_begin_time) {
-    cycle_begin_time_ = cycle_begin_time;
-    try {
-        // Get current motion
-        Motion current_motion = getCurrentMotion();
-        // If the new LTT was processed at least once and is labeled safe, replace old LTT with new one.
-        if (new_ltt_ && new_ltt_processed_) {
-            if ((is_static_safe_ && is_PFL_safe_) || current_motion.isStopped()) {
-                long_term_trajectory_ = new_long_term_trajectory_;
-                new_ltt_ = false;
-                new_goal_ = false;
-            }
-        }
-        // Check if there is a new goal motion
-        if (new_goal_) {
-            // Check if current motion has acceleration and jerk values that lie in the plannable ranges
-            bool is_plannable = checkCurrentMotionForReplanning(current_motion);
-            if (is_plannable) {
-                // Check if the starting position of the last replanning was very close to the current position
-                bool last_close = true;
-                if (new_ltt_ == true) {
-                    for (int i = 0; i < current_motion.getAngle().size(); i++) {
-                        if (std::abs(current_motion.getAngle()[i] - last_replan_start_motion_.getAngle()[i]) > 0.01) {
-                            last_close = false;
-                            break;
-                        }
-                    }
-                } else {
-                    last_close = false;
-                }
-                // Only replan if the current joint position is different from the last.
-                bool success = true;
-                if (!last_close) {
-                    success = calculateLongTermTrajectory(current_motion.getAngle(), current_motion.getVelocity(),
-                                                          current_motion.getAcceleration(), new_goal_motion_.getAngle(), new_long_term_trajectory_);
-                    if (success) {
-                        last_replan_start_motion_ = current_motion;
-                    }
-                }
-                if (success) {
-                    new_ltt_ = true;
-                    new_ltt_processed_ = false;
-                } else {
-                    new_ltt_ = false;
-                }
-            } else {
-                new_ltt_ = false;
-            }
-        }
-        // If there is a new long term trajectory (LTT), always override is_safe with false.
-        if (new_ltt_ && !new_ltt_processed_) {
-            is_PFL_safe_ = false;
-            is_static_safe_ = false;
-        }
-        // Compute a new potential trajectory
-        Motion goal_motion_static;
-        Motion goal_motion_pfl;
-        computesPotentialTrajectoryForPFL(is_static_safe_, is_PFL_safe_, next_motion_.getVelocity(), &goal_motion_static, &goal_motion_pfl);
-        if (activate_shield_) {
-            // Check motion for joint limits
-            bool pfl_joint_limit = true;
-            if(!is_under_iso_velocity_) {
-                pfl_joint_limit = checkMotionForJointLimits(goal_motion_pfl);
-            }
-            // TODO: very conservative, could improve this depending on verification?
-            if (!(checkMotionForJointLimits(goal_motion_static) && pfl_joint_limit)) {
-                is_static_safe_ = false;
-                is_PFL_safe_ = false;
-            } else {
-                // check if robot doesnt run into static human
-                robot_capsules_static_ = robot_reach_->reach(current_motion, goal_motion_static, (goal_motion_static.getS()-current_motion.getS()), alpha_i_);
-                // TODO: humanReach, immer die kleinste Kapsel ist wichtig
-                human_reach_->humanReachabilityAnalysis(cycle_begin_time_, 0); //t_brake is differentiell
-                human_capsules_static_ = human_reach_->getAllCapsules();
-                is_static_safe_ = verify_->verify_human_reach(robot_capsules_static_, human_capsules_static_);
-
-                // check if there is collision between now and time t, during that time velocity of robot is higher than v_iso_
-                // velocity_criteria is true, when v_max <= v_iso
-                if(is_under_iso_velocity_) {
-                    is_PFL_safe_ = true;
-                    // if v_max <= v_iso, no velocity capsules get calculated because velocity criterion is true
-                    // so we set the velocity capsules to the static capsules
-                    robot_capsules_velocity_ = robot_capsules_static_;
-                    human_capsules_velocity_ = human_capsules_static_;
-                } else {
-                    robot_capsules_velocity_ = robot_reach_->reach(current_motion, goal_motion_pfl, (goal_motion_pfl.getS()-current_motion.getS()), alpha_i_);
-                    human_reach_->humanReachabilityAnalysis(cycle_begin_time_, goal_motion_pfl.getTime());
-                    human_capsules_velocity_ = human_reach_->getAllCapsules();
-                    is_PFL_safe_ = verify_->verify_human_reach(robot_capsules_velocity_, human_capsules_velocity_);
-                }
-
-            }
-        } else {
-            is_static_safe_ = true;
-            is_PFL_safe_ = true;
-        }
-        // Select the next motion based on the verified safety
-        next_motion_ = determineNextMotionForPFL(is_static_safe_, is_PFL_safe_);
-        next_motion_.setTime(cycle_begin_time);
-        new_ltt_processed_ = true;
-        return next_motion_;
-    } catch (const std::exception &exc) {
-        spdlog::error("Exception in SafetyShield::getNextCyclePFL: {}", exc.what());
-    }
-}
-
-Motion SafetyShield::getCurrentMotion() {
-    Motion current_pos;
-    if(safety_method_ == STANDARD) {
-        if (!recovery_path_.isCurrent()) {
-            current_pos = long_term_trajectory_.interpolate(failsafe_path_.getPosition(),
-                                                            failsafe_path_.getVelocity(),
-                                                            failsafe_path_.getAcceleration(),
-                                                            failsafe_path_.getJerk(),
-                                                            v_max_allowed_,
-                                                            a_max_allowed_,
-                                                            j_max_allowed_);
-        } else {
-            current_pos = long_term_trajectory_.interpolate(recovery_path_.getPosition(),
-                                                            recovery_path_.getVelocity(),
-                                                            recovery_path_.getAcceleration(),
-                                                            recovery_path_.getJerk(),
-                                                            v_max_allowed_,
-                                                            a_max_allowed_,
-                                                            j_max_allowed_);
-        }
-        return current_pos;
-    } else {
-        if (failsafe_path_static_.isCurrent()) {
-            current_pos = long_term_trajectory_.interpolate(failsafe_path_static_.getPosition(),
-                                                            failsafe_path_static_.getVelocity(),
-                                                            failsafe_path_static_.getAcceleration(),
-                                                            failsafe_path_static_.getJerk(),
-                                                            v_max_allowed_,
-                                                            a_max_allowed_,
-                                                            j_max_allowed_);
-        } else if(failsafe_path_pfl_.isCurrent()) {
-            current_pos = long_term_trajectory_.interpolate(failsafe_path_pfl_.getPosition(),
-                                                            failsafe_path_pfl_.getVelocity(),
-                                                            failsafe_path_pfl_.getAcceleration(),
-                                                            failsafe_path_pfl_.getJerk(),
-                                                            v_max_allowed_,
-                                                            a_max_allowed_,
-                                                            j_max_allowed_);
-        } else {
-            if(!recovery_path_.isCurrent()) {
-                spdlog::error("getCurrentMotion: no path is set?");
-            }
-            current_pos = long_term_trajectory_.interpolate(recovery_path_.getPosition(),
-                                                            recovery_path_.getVelocity(),
-                                                            recovery_path_.getAcceleration(),
-                                                            recovery_path_.getJerk(),
-                                                            v_max_allowed_,
-                                                            a_max_allowed_,
-                                                            j_max_allowed_);
-        }
-        return current_pos;
-    }
-}
-
-bool SafetyShield::checkCurrentMotionForReplanning(Motion& current_motion) {
-    for (int i = 0; i < nb_joints_; i++) {
-        if (std::abs(current_motion.getAcceleration()[i]) > a_max_ltt_[i]) {
-            return false;
-        }
-    }
-    return true;
-}
-
-void SafetyShield::newLongTermTrajectory(const std::vector<double> &goal_position,
-                                         const std::vector<double> &goal_velocity) {
-    try {
-        std::vector<double> motion_q;
-        motion_q.reserve(nb_joints_);
-        std::vector<double> motion_dq;
-        motion_dq.reserve(nb_joints_);
-        for (int i = 0; i < nb_joints_; i++) {
-            // TODO: replace with config max min values
-            motion_q.push_back(std::clamp(goal_position[i], q_min_allowed_[i], q_max_allowed_[i]));
-            motion_dq.push_back(std::clamp(goal_velocity[i], -v_max_allowed_[i], v_max_allowed_[i]));
-        }
-        new_goal_motion_ = Motion(cycle_begin_time_, motion_q, motion_dq);
-        new_goal_ = true;
-        new_ltt_ = false;
-        new_ltt_processed_ = false;
-        std::vector<double> dummy_q;
-        for (int i = 0; i < nb_joints_; i++) {
-            // fill with high values to guarantee this is not close to current joint position
-            dummy_q.push_back(-1234567);
-        }
-        last_replan_start_motion_ = Motion(cycle_begin_time_, dummy_q);
-    } catch (const std::exception &exc) {
-        spdlog::error("Exception in SafetyShield::newLongTermTrajectory: {}", exc.what());
-    }
-}
-
-void SafetyShield::setLongTermTrajectory(LongTermTraj& traj) {
-    Motion current_motion = getCurrentMotion();
-    // Check if robot is at stop
-    if (!current_motion.isStopped()) {
-        throw RobotMovementException();
-    }
-    Motion start = traj.getNextMotionAtIndex(0);
-    // Check if traj starts at the same position
-    if (!current_motion.hasSamePos(&start)) {
-        throw TrajectoryException("Given LTT does not start at current robot position.");
-    }
-    // Check if traj starts at v=0
-    if (!current_motion.hasSameVel(&start)) {
-        std::stringstream ss;
-        ss << "Given LTT does not start with velocity 0.0. Start velocity is [";
-        for(size_t i = 0; i < start.getVelocity().size(); ++i)
-        {
-            if(i != 0)
-                ss << ",";
-            ss << start.getVelocity()[i];
-        }
-        ss << "].";
-        std::string s = ss.str();
-        throw TrajectoryException(s);
-    }
-    // Check if traj ends in stop
-    if (!traj.getNextMotionAtIndex(traj.getLength()-1).isStopped()) {
-        throw TrajectoryException("Given LTT does not end in a complete stop of the robot (v = a = j = 0.0)");
-    }
-    // Replace LTT
-    new_long_term_trajectory_ = traj;
-    new_ltt_ = true;
-    new_ltt_processed_ = true;
-}
-
-bool SafetyShield::calculateLongTermTrajectory(const std::vector<double>& start_q,
-                                               const std::vector<double> start_dq,
-                                               const std::vector<double> start_ddq,
-                                               const std::vector<double>& goal_q,
-                                               LongTermTraj& ltt) {
-    long_term_planner::Trajectory trajectory;
-    bool success = ltp_.planTrajectory(goal_q, start_q, start_dq, start_ddq, trajectory);
-    if (!success) return false;
-    std::vector<Motion> new_traj(trajectory.length);
-    double new_time = path_s_;
-    std::vector<double> q(nb_joints_);
-    std::vector<double> dq(nb_joints_);
-    std::vector<double> ddq(nb_joints_);
-    std::vector<double> dddq(nb_joints_);
-    for (int i = 0; i < trajectory.length; i++) {
-        for (int j=0; j < nb_joints_; j++) {
-            q[j] = trajectory.q[j][i];
-            dq[j] = trajectory.v[j][i];
-            ddq[j] = trajectory.a[j][i];
-            dddq[j] = trajectory.j[j][i];
-        }
-        new_traj[i] = Motion(new_time, q, dq, ddq, dddq);
-        new_time += sample_time_;
-    }
-    if(safety_method_ == STANDARD) {
-        ltt = LongTermTraj(new_traj, sample_time_, path_s_discrete_, sliding_window_k_);
-    } else {
-        ltt = LongTermTraj(new_traj, sample_time_, *robot_reach_, path_s_discrete_, sliding_window_k_);
-    }
-    return true;
-}
-
-} // namespace safety_shield
